Browser Router added links work
