+++ src/components/Navbar.js
@@ -1,4 +1,10 @@
 import React from 'react';
+import {
+  BrowserRouter as Router,
+  Switch,
+  Route,
+  Link
+} from "react-router-dom";
 
 const Navbar = () => {
   return(
@@ -6,17 +12,19 @@ const Navbar = () => {
       <div className="navbar_center">
         <img src={"https://static.files.bbci.co.uk/orbit/0b0a71ff229422fc63da1ec4b9726123/img/blq-orbit-blocks_white_alpha.png"} />
         <ul className="navbar_menu">
-          <li>Sign in</li>
-          <li>Home</li>
-          <li>News</li>
-          <li>Sport</li>
-          <li>Reel</li>
-          <li>Worklife</li>
-          <li>Travel</li>
-          <li>Future</li>
-          <li>Culture</li>
-          {/* Dumme for Dropdown */}
-          <li>More</li>
+        <Router>
+            <li><Link to="#">Sign in</Link></li>
+            <li><Link to="#">Home</Link></li>
+            <li><Link to="#">News</Link></li>
+            <li><Link to="#">Sport</Link></li>
+            <li><Link to="#">Reel</Link></li>
+            <li><Link to="#">Worklife</Link></li>
+            <li><Link to="#">Travel</Link></li>
+            <li><Link to="#">Future</Link></li>
+            <li><Link to="#">Culture</Link></li>
+            {/* Dumme for Dropdown */}
+            <li><Link to="#">More</Link></li>
+          </Router>
         </ul>
       </div>
     </div>
